refactor: replace int with size_t where appropriate

+++ src/CrossMappingCardinality.cpp
@@ -279,7 +279,7 @@ std::vector<std::vector<double>> CrossMappingCardinality2(
   // // Sequential version of the for loop
   // if (progressbar) {
   //   RcppThread::ProgressBar bar(num_neighbors, 1);
-  //   for (int i = 0; i < num_neighbors; ++i) {
+  //   for (size_t i = 0; i < num_neighbors; ++i) {
   //     std::vector<double> H1sliced(H1sequence.begin(), H1sequence.begin() + i + 1);
   //     std::vector<double> dp_res = CppCMCTest(H1sliced,">",0.05,num_neighbors);
   //     dp_res.insert(dp_res.begin(), i + 1);
@@ -287,7 +287,7 @@ std::vector<std::vector<double>> CrossMappingCardinality2(
   //     bar++;
   //   };
   // } else {
-  //   for (int i = 0; i < num_neighbors; ++i) {
+  //   for (size_t i = 0; i < num_neighbors; ++i) {
   //     std::vector<double> H1sliced(H1sequence.begin(), H1sequence.begin() + i + 1);
   //     std::vector<double> dp_res = CppCMCTest(H1sliced,">",0.05,num_neighbors);
   //     dp_res.insert(dp_res.begin(), i + 1);
@@ -298,7 +298,7 @@ std::vector<std::vector<double>> CrossMappingCardinality2(
   // Parallel computation with or without a progress bar
   if (progressbar) {
     RcppThread::ProgressBar bar(num_neighbors, 1);
-    RcppThread::parallelFor(0, num_neighbors, [&](int i) {
+    RcppThread::parallelFor(0, num_neighbors, [&](size_t i) {
       std::vector<double> H1sliced(H1sequence.begin(), H1sequence.begin() + i + 1);
       std::vector<double> dp_res = CppCMCTest(H1sliced,">",0.05,num_neighbors);
       dp_res.insert(dp_res.begin(), i + 1);
@@ -306,7 +306,7 @@ std::vector<std::vector<double>> CrossMappingCardinality2(
       bar++;
     }, threads_sizet);
   } else {
-    RcppThread::parallelFor(0, num_neighbors, [&](int i) {
+    RcppThread::parallelFor(0, num_neighbors, [&](size_t i) {
       std::vector<double> H1sliced(H1sequence.begin(), H1sequence.begin() + i + 1);
       std::vector<double> dp_res = CppCMCTest(H1sliced,">",0.05,num_neighbors);
       dp_res.insert(dp_res.begin(), i + 1);
